API/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.66.3
